Relocate record nft mint
